Move test_aligned_vec
